documentation is revised
Also, `maximum_capacity` is added to debug information.
