+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "orx-concurrent-bag"
-version = "1.8.1"
+version = "1.8.2"
 edition = "2021"
 authors = ["orxfun <orx.ugur.arikan@gmail.com>"]
 description = "An efficient, convenient and lightweight grow-only concurrent collection."
@@ -12,7 +12,7 @@ categories = ["data-structures", "concurrency", "rust-patterns"]
 [dependencies]
 orx-fixed-vec = "2.6.1"
 orx-pinned-vec = "2.6"
-orx-split-vec = "2.7"
+orx-split-vec = "2.7.1"
 
 
 [dev-dependencies]
