+++ src/common_traits/debug.rs
@@ -10,6 +10,29 @@ impl<T: Debug, P: PinnedVec<T>> Debug for ConcurrentBag<T, P> {
             .field("pinned", &unsafe { self.iter().collect::<Vec<_>>() })
             .field("len", &self.len())
             .field("capacity", &self.capacity())
+            .field("maximum_capacity", &self.maximum_capacity())
             .finish()
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn debug() {
+        let bag = ConcurrentBag::new();
+
+        bag.push('a');
+        bag.push('b');
+        bag.push('c');
+        bag.push('d');
+        bag.push('e');
+
+        let str = format!("{:?}", bag);
+        assert_eq!(
+            str,
+            "ConcurrentBag { pinned: ['a', 'b', 'c', 'd', 'e'], len: 5, capacity: 12, maximum_capacity: 17179869180 }"
+        );
     }
 }
